Replaced float and int with CGFloat and NSInteger respectively

+++ VRGCalendarView/VRGCalendarView.h
@@ -42,10 +42,10 @@
 @property (nonatomic, retain) UIImageView *animationView_B;
 @property (nonatomic, retain) NSArray *markedDates;
 @property (nonatomic, retain) NSArray *markedColors;
-@property (nonatomic, getter = calendarHeight) float calendarHeight;
+@property (nonatomic, getter = calendarHeight) CGFloat calendarHeight;
 @property (nonatomic, retain, getter = selectedDate) NSDate *selectedDate;
 
--(void)selectDate:(int)date;
+-(void)selectDate:(NSInteger)date;
 -(void)reset;
 
 -(void)markDates:(NSArray *)dates;
@@ -54,13 +54,13 @@
 -(void)showNextMonth;
 -(void)showPreviousMonth;
 
--(int)numRows;
+-(NSInteger)numRows;
 -(void)updateSize;
 -(UIImage *)drawCurrentState;
 
 @end
 
 @protocol VRGCalendarViewDelegate <NSObject>
--(void)calendarView:(VRGCalendarView *)calendarView switchedToMonth:(int)month year:(int)year numOfDays:(int)days targetHeight:(float)targetHeight animated:(BOOL)animated;
+-(void)calendarView:(VRGCalendarView *)calendarView switchedToMonth:(NSInteger)month year:(NSInteger)year numOfDays:(NSInteger)days targetHeight:(CGFloat)targetHeight animated:(BOOL)animated;
 -(void)calendarView:(VRGCalendarView *)calendarView dateSelected:(NSDate *)date;
 @end
